requirements.txt
mcp>=0.1.0
tushare>=2.0.0
python-dotenv>=0.19.0
pandas>=1.5.0 